Use NSStackView for find window’s status bar and move to view controller
The main advantage is that we can let the stack view handle showing and hiding of the stop button and progress indicator.

The height of the status bar (text field) is now fixed at 16 pt. The reason we have to fix it is that we are using a button for the status bar (so that the user can click toggle between showing number of files and number of bytes searched), but a button with an empty string reports a different fittingSize than one with a non-empty string.

Therefore without a fixed height, we need to ensure status bar never gets set to an empty string, which we did before, but I think it is nicer to simply fix the height of the status bar.

Also because the progress indicator and stop buttons might actually be slightly taller than the status bar used to be.