Improve performance when showing clipboard history
Previously we fetched items row-by-row from the database, which could be slow even with only 500 entries.

We now fetch everything using a single query. Long-term though we may fetch the strings on-demand but with some prefetch, mainly just to avoid an issue if the user store hundred of megabytes in the clipboard history.