Let our table cell view observe all outline views for looser coupling
The motivation is not having to know about the interface of the view controller that creates the instance.

It does have performance disadvantages in that collapsing or expanding items in *any* outline view will now trigger notification callbacks for our table header cell views.

In practice though expanding/collapsing is not a common action, and the chance of this being done, while the find window has so many results as to impact performance, is probably small (it’s only the header cells that register for this notification).

That said, it would be good to find a better solution.