Introduce FFTextFieldViewController
This encapsulates syntax highlight, history, and auto-resize behavior.

We currently do not call updateIntrinsicContentSizeToEncompassString: when changing the content, but it appears to be unnecessary. It might actually be that with auto-layout, there is no longer any need for repeatedly updating the intrinsic height of the control manually.

But leaving things (more or less) as-is for now.