Set y-placement to ‘none’ for labels in preferences window
Rows are baseline aligned, and when that is the case, the grid view will use the y-placement of the first cell (which is not ‘none’) to place the row.

However, since the non-label column is likely to contain taller views, we should use that for placing the row.

That said, this change has no effect on the resulting layout, possibly because there are constraints with higher priorities that prevent the labels from bleeding into neighboring rows.