Use custom code for preferences window
Relying on “modern” APIs we can eliminate most of the MASPreferencesViewController protocol:

- viewIdentifier → identifier
- toolbarItemLabel → title
- initialKeyView → call makeFirstResponder: in viewDidAppear
- hasResizableWidth / hasResizableHeight → use auto layout constraints

Furthermore, handling the preferences panes can be done by using child view controllers, and it might even be possible to use transitionFromViewController:toViewController:options:completionHandler: to switch between the views.

The last point is actually the motivation for this change, as I need a playground for animated view transitions.