Avoid potential crash by not updating window’s opaque property
This does completely remove support for transparent themes, but any recent version of macOS would not support it anyway due to layer-backed views.

The crash resulting from setting opaque is because of recursion: After session restore, the window will calculate shadow, this triggers a “user defaults did change” notification, which updates theme of text view, which sets opaque, which re-calculates shadow and tries to recursively obtain a non-recursive mutex.