Use single quotes when using ‘notarize_pass’ and ‘notarize_user’
These variables are passed as arguments to commands that go through a shell interpreter, therefore spaces would be a problem.

Previously we expected the variables themselves to include proper escaping (or quotes), but that is problematic if they are set in build.ninja, as we recreate this file via a rule that passes all variables as arguments, and this rule already uses single quotes for escaping, so it would cause a problem, if the variables contains single quotes.