Add build rule that expands variables in InfoPlist.vstrings files
The extension is converted to from ‘vstrings’ to ‘strings’ and it is saved as UTF-16. Ideally this would not introduce a new extension (and rely on the existing ‘strings’ filter to convert to UTF-16), but see previous commit for the technical limitation preventing this.

TARGET_NAME and YEAR are predefined variables.