build: Do not disable filters after first “generator”
Previously we would only allow a filter when going from source to build directory, since the output base name is unchanged for a filter’s output, and would thus cause multiple targets with same output name.

Since we now add the generator’s name to the output path, this is no longer a problem.