Only use atomic save for remote volumes (by default)
APFS does not support exchangedata(2) meaning that atomic save has to be done using a rename.

This means that the inode is not preserved and the date of the file’s directory is updated.

There are some disadvantages to this, and I have been running with atomic saving disabled shortly after switching to APFS, so I believe it is best to make this the default behavior.