Fix truncated text in tool tips using Dark Mode on macOS 10.15
We add a border to the NSTextField in Dark Mode to match the system tool tips, but it appears that the frame returned by “fittingSize” does not take that into account, so it returns a frame 1-2 points too small to fit the text, resulting in the last line missing.

This appears to be a fairly new problem, so might depend on the SDK that we link with.