Rave: Treat all values with a ‘${dir}/’ or ‘${HOME}/’ prefix as files
Previously we used a heuristic and if the value looked like a path, checked that it actually existed.

The new approach makes the rules simpler, avoids a file system check, and also ensures that non-existing files *will* become dependencies.

The latter is useful if they are generated by local.ninja, typos, or actually unmet dependencies.