Change HTML output window back from NSPanel to NSWindow
The problem is that if another panel opens, that panel is not tied to our HTML output window, for example opening the Find dialog does not allow us to press return to search in the HTML output, and if we open a modal dialog, when that modal dialog is disposed, the *main window* will be brought to front (not the previously focused panel).

This effectively reverts commit 92953099e6230cd61c921039084185eab0aac7b7.