Rave: Fix issue with generated headers and universal builds
Generated Ragel and Cap’n Proto sources are required by both Intel and Arm targets, but as the output is the same, we cache and re-use the result from these transformations.

This however meant that our heuristic to find generated headers, and add these to the include path, would fail when we are receiving cached results (as we skip the intermediate, identical, steps).

We now return the intermediate steps from the cache, but mark them as duplicates, so that they can be stripped when generating the build file.

Not an ideal solution, but the real issue is really identifying generated headers, and getting these added to the include path.