Rework API for the Open With submenu
The API expose a few more implementation details and require the user retains the delegate for the duration of the menu’s lifecycle.

The upside is that we avoid the global singleton and it is now possible to examine and use the information gathered by the Open With menu delegate.