OakTabBarViewController: No implicit animation when re-ordering tabs
This is only relevant when using the binding interface, instead of manual reload.

The tab bar cannot do an animated relayout when tabs are re-ordered. It probably should detect this itself, and skip the animation, though the animation group is setup in the controller, so this would still cause some side-effects.