Increase padding in front of pop-ups in text view status bar
Prior to Big Sur, the internal left-padding on borderless NSPopUpButtons was excessive, which is why we used negative space to limit the amount of whitespace.

It appears that the default spacing is now more appropriate (on Big Sur).

Ideally we would only increase the spacing on Big Sur, but the extra spacing doesn’t look too bad on Catalina (and presumably earlier), so I am going with the Big Sur spacing constants for all builds, to keep the code simpler.