Increase padding for file browser’s top folder pop-up
The spacing was so that the icon (in the folder pop-up) was aligned with the file item icons in the list below, but this no longer holds on Big Sur, and the spacing is too tight on Big Sur.

Like with the status bar pop-ups, I am using the increased spacing for all builds, to keep things simple.