Ensure low content hugging priority for text field used in table view
When the NSTextField is not editable then it will (by default) hug its content, which is undesired, since the potential views following the text field (tags and close button) should appear right-aligned.

In general though, we do not have controls after the text field when it is read-only, as it represents a missing file, so this is just for layout consistency.