Remove deployment target from all xib files
We pass --minimum-deployment-target to ibtool so having the deployment target in the xib is redundant.

Furthermore, it appears that when building on macOS 11, a directory will be created for xib files when they:

1. Are a custom view (as opposed to a window)
2. Contains a deployment target
3. Contains a table view

Therefore, to successfully build on macOS 11 we cannot set deployment target for xib files meeting the above 3 criteria, which is easier when we only set the target in one place, instead of every single xib.

Thanks to Kevin Sjöberg for helping track this down.