Allow creating TMFileReference from an image instead of URL
This is for places where we normally want to show a (live) image representing a URL (and its modified/SCM status) via bindings, but occasionally, we want to show a static image.

For example the location shown in the file browser. This is normally a folder, but it could be “Computer”. Similar for folder search results when searching unsaved documents.