Change default size of image returned by TMFileReference: 48×48 → 16×16
This makes binding menu items to a file reference image much simpler, as we do not have to bind to a property wrapper that resizes the image.

As a general rule, we always copy an image, before we change its size, so the image obtained from TMFileReference should always stay at its default size.